extension Int: Add type cast.

+++ Concentration/ViewController.swift
@@ -75,7 +75,7 @@ extension Int {
         if self > 0 {
             return Int(arc4random_uniform(UInt32(self)))
         } else if self < 0 {
-            return -Int(arc4random_uniform(UInt32(abs(self))))
+            return -Int(arc4random_uniform(UInt32(abs((Int32(self))))))
         } else {
             return 0
         }
